updated async call

+++ app/src/main/java/com/recipewizard/recipewizard/Ingredient.java
@@ -9,13 +9,13 @@ import android.os.Parcel;
  * Created by Brian on 11/17/2016.
  */
 
-public class Ingredient implements Parcelable{
+public class Ingredient implements Parcelable {
 
-    private String name = new String ();
+    private String name = new String();
     private Bitmap picture;
     private boolean checked;
     private String category = new String();
-    private int amount;
+    private double amount;
     private String unit;
 
     Ingredient(String name, Bitmap picture, boolean checked, String category){
@@ -23,7 +23,7 @@ public class Ingredient implements Parcelable{
         this.picture = picture;
         this.checked = checked;
     }
-    Ingredient(String name, Bitmap picture, boolean checked, String category, int amount,
+    Ingredient(String name, Bitmap picture, boolean checked, String category, double amount,
                String unit){
         this.name = name;
         this.picture = picture;
@@ -53,8 +53,8 @@ public class Ingredient implements Parcelable{
     }
     public String getCategory() {return category;}
     public void setCategory(String category) {this.category = category;}
-    public int getAmount() {return amount;}
-    public void setAmount(int amount) {this.amount = amount;}
+    public double getAmount() {return amount;}
+    public void setAmount(double amount) {this.amount = amount;}
     public String getUnit () {return unit;}
     public void setUnit (String unit) {this.unit = unit;}
 
@@ -73,7 +73,7 @@ public class Ingredient implements Parcelable{
         dest.writeValue(picture);
         dest.writeByte((byte) (checked ? 1 : 0));
         dest.writeString(category);
-        dest.writeInt(amount);
+        dest.writeDouble(amount);
         dest.writeString(unit);
     }
